moved generate and helper functions to gameutils/setup
